requirements.txt
customtkinter>=5.2.0
pandas>=2.2.2
openpyxl>=3.1.2
zeep>=4.2.1
python-docx>=1.1.0
docx2pdf>=0.1.8